refactor project
create classes for evil files

create examples and HowTo md files
