+++ HowTo/Malwares/SendEmail.md
@@ -1,0 +1,22 @@
+# Send Email Requirements
+
+## Steps
+
+- Turn on 2FA
+- Go to GoogleAcc->Security->AppPassword->Custom
+- Copy generated key and use that as password
+
+## Use malwares.utils module
+
+```python
+from pyhtools.malwares.utils import send_mail
+
+send_mail(
+    email = 'email.id@domain.com', 
+    password = 'your',
+    receiver_mail = 'emails_separated_by_comma',
+    message = 'your message',
+    smtp_server = 'smtp.gmail.com',
+    smtp_port= 587
+)
+```
