+++ HowTo/Malwares/TelegramDataHarvester.md
@@ -1,0 +1,18 @@
+# How to use Telegram Data Harvester
+
+## Steps
+
+- Create a Gmail App Password
+- update telegram_data_harvester.py lines 64 and 65 with your gmail and App Password
+- Create a executable file or trojan
+- Send to the target
+- You will receive an email with telegram data as a zip when target executes the program/trojan.
+
+## How to use their account
+
+- Extract tdata folder from the zip file
+- go to your installed telegram desktop location
+- replace your tdata file with the zip tdata folder
+- Start Telegram Desktop
+
+> Tested on Windows 10 and Ubuntu
