+++ pyhtools/evil_files/malwares/utils.py
@@ -1,0 +1,36 @@
+import smtplib
+import requests
+
+
+def send_mail(email: str, receiver_mail: list[str], password: str, message: str, smtp_server: str = 'smtp.gmail.com', smtp_port: int = 587) -> bool:
+    '''
+    sends mail to specific address/addresses.
+    '''
+    try:
+        server = smtplib.SMTP(smtp_server, smtp_port)
+        server.starttls()
+        server.login(email, password)
+        server.sendmail(email, receiver_mail, message)
+        server.quit()
+        return True
+    except smtplib.SMTPException as e:
+        print('[-] Exception : ', e)
+        return False
+
+
+def download(url: str) -> bool:
+    '''
+    description: download and save file from the url.
+    params: url
+    returns: bool
+    '''
+    try:
+        response = requests.get(url)
+        file_name = url.split('/')[-1]
+        with open(file_name, 'wb') as file:
+            file.write(response.content)
+        return True
+
+    except Exception as e:
+        print('[-] Exception : ', e)
+        return False
