--- pyhtools/executable_generator/utils.py
@@ -1,2 +1,0 @@
-from pyhtools.UI.colors import *
-
