--- pyhtools/malwares/TelegramRemoteCodeExecutor/HowToUse.md
@@ -1,17 +1,0 @@
-# How to use Telegram Remote Code Executor
-
-- Get Token for bot from [Telegram BotFather](https://t.me/botfather) and update `API_KEY` on line 7 of `TelegramRemoteCodeExecutor.py` 
-- Get Attacker's User ID from [UserInfoBot](https://t.me/userinfobot) or run TelegramRemoteCodeExecutor.py and use `/start` command to get user id and then update `CHAT_ID` on line 8 of `TelegramRemoteCodeExecutor.py` 
-- Create trojan/malware using `pyinstaller` and send it to victim
-- When victims runs malware program, attacker will be notified through the bot on telegram
-- Attacker can use commands to exploit the machine remotely through Telegram over the internet
-
-
-| Command |  Description | Params |
-|:-----:|:--------:|:-------: |
-| /start   | starts interaction with the bot and returns command issuer details  | None |
-| /help   |  returns help menu  | None |
-| /exec   | executes command on vitcim's machine | command |
-| /cd   | change current working directory to specified path | path |
-| /ls | list all the directories and folders in the current working directory | None |
-| /download | download file from the victims machine to attacker's via telegram chat | relative or complete file path |
