--- pyhtools/malwares/downloader/downloader.py
@@ -1,18 +1,0 @@
-import requests
-
-def download(url:str)->bool:
-    '''
-    description: download and save file from the url.
-    params: url
-    returns: bool
-    '''
-    try:
-        response = requests.get(url)
-        file_name = url.split('/')[-1]
-        with open(file_name, 'wb') as file:
-            file.write(response.content)
-        return True
-
-    except Exception as e:
-        print('[-] Exception : ', e)
-        return False
