--- pyhtools/malwares/send_mail/how to.txt
@@ -1,6 +1,0 @@
-refer this video tutorial: https://www.youtube.com/watch?v=ql5Dex4m40w
-
-Steps:
-1. Turn on 2FA
-2. Go to GoogleAcc->Security->AppPassword->Custom
-3. Copy generated key and use that as password
