--- pyhtools/malwares/wireless_password_harvester/HowToUse.txt
@@ -1,4 +1,0 @@
-HOW to use:
-1. Update email id and password
-2. Turn on less secure app: https://myaccount.google.com/lesssecureapps
-3. Create exe: pyinstaller.exe --onefile --icon='icon_path' -n 'name_of_file' .\wireless_profile_harvestor.py
